+++ packages/component/src/ScrollToBottom/State2Context.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const context = React.createContext({
+  animating: false,
+  animatingToEnd: false,
+  sticky: true
+});
+
+context.displayName = 'ScrollToBottomState2Context';
+
+export default context;