+++ packages/component/src/hooks/useAtStart.js
@@ -1,7 +1,7 @@
 import useStateContext from './internal/useStateContext';
 
 export default function useAtStart() {
-  const { atStart } = useStateContext();
+  const { atStart } = useStateContext(1);
 
   return [atStart];
 }