+++ packages/component/src/hooks/useAtTop.js
@@ -1,7 +1,7 @@
 import useStateContext from './internal/useStateContext';
 
 export default function useAtTop() {
-  const { atTop } = useStateContext();
+  const { atTop } = useStateContext(1);
 
   return [atTop];
 }